Revert "Fix -Wimplicit-int-float-conversion."
This reverts commit 7ec6dc28b4de77edbcef90af51a19af0af28e45c.

Reason for revert: This is breaking iOS builds: 
https://ci.chromium.org/ui/p/chromium/builders/try/ios-simulator-cronet/48182/overview

Original change's description:
> Fix -Wimplicit-int-float-conversion.
>
> Bug: 989932
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3001416
> Commit-Queue: Peter Kasting <pkasting@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#901649}

Bug: 989932
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3030565
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Commit-Queue: Christian Dullweber <dullweber@chromium.org>
Owners-Override: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901936}

+++ media/gpu/test/image_quality_metrics.cc
@@ -138,12 +138,11 @@ double ComputeYUV420P10SSIM(const uint8_t* src_y_a,
 double ComputeSimilarity(const VideoFrame* frame1,
                          const VideoFrame* frame2,
                          SimilarityMetrics mode) {
-  ASSERT_TRUE_OR_RETURN(
-      frame1->IsMappable() && frame2->IsMappable(),
-      static_cast<double>(std::numeric_limits<std::size_t>::max()));
+  ASSERT_TRUE_OR_RETURN(frame1->IsMappable() && frame2->IsMappable(),
+                        std::numeric_limits<std::size_t>::max());
   ASSERT_TRUE_OR_RETURN(
       frame1->visible_rect().size() == frame2->visible_rect().size(),
-      static_cast<double>(std::numeric_limits<std::size_t>::max()));
+      std::numeric_limits<std::size_t>::max());
   // Ideally, frame1->BitDepth() should be the same as frame2->BitDepth()
   // always. But in the 10 bit case, the 10 bit frame can be carried with P016LE
   // whose bit depth is regarded to be 16. This is due to a lack of NV12 10-bit
@@ -152,7 +151,7 @@ double ComputeSimilarity(const VideoFrame* frame1,
   ASSERT_TRUE_OR_RETURN(
       (frame1->BitDepth() == 8 && frame1->BitDepth() == frame2->BitDepth()) ||
           std::min(frame1->BitDepth(), frame2->BitDepth()) == 10,
-      static_cast<double>(std::numeric_limits<std::size_t>::max()));
+      std::numeric_limits<std::size_t>::max());
   const size_t bit_depth = std::min(frame1->BitDepth(), frame2->BitDepth());
   const VideoPixelFormat common_format =
       bit_depth == 8 ? PIXEL_FORMAT_I420 : PIXEL_FORMAT_YUV420P10;
